Initial Feeding System
Implemented drag/drop function for feeding a character and context-sensitive cursor mappings. Food UI is pulled from inventory and can be dragged to a trigger on the character to eat it, which deducts it from the inventory and updates the food/qty text.

Selection can be changed via button or horizontal input axis.